moved elements to base page
