+++ src/app/shared/classes/test-flow-base-pages/pass-finalisation/pass-finalisation-base-page.ts
@@ -4,16 +4,64 @@ import { Platform } from '@ionic/angular';
 import { Router } from '@angular/router';
 
 import { StoreModel } from '@shared/models/store.model';
-import { getCurrentTest, getJournalData } from '@store/tests/tests.selector';
+import { getCurrentTest, getJournalData, getTestOutcomeText } from '@store/tests/tests.selector';
 import { getTests } from '@store/tests/tests.reducer';
 import { AuthenticationProvider } from '@providers/authentication/authentication';
 import { getCandidate } from '@store/tests/journal-data/common/candidate/candidate.reducer';
-import { getCandidateName } from '@store/tests/journal-data/common/candidate/candidate.selector';
+import {
+  formatDriverNumber,
+  getCandidateDriverNumber,
+  getCandidateName,
+  getUntitledCandidateName,
+} from '@store/tests/journal-data/common/candidate/candidate.selector';
 
 import { PracticeableBasePageComponent } from '@shared/classes/practiceable-base-page';
+import { map } from 'rxjs/operators';
+import { getApplicationReference }
+  from '@store/tests/journal-data/common/application-reference/application-reference.reducer';
+import { getApplicationNumber }
+  from '@store/tests/journal-data/common/application-reference/application-reference.selector';
+import { getPassCompletion } from '@store/tests/pass-completion/pass-completion.reducer';
+import {
+  getPassCertificateNumber,
+  isProvisionalLicenseProvided,
+} from '@store/tests/pass-completion/pass-completion.selector';
+import { getVehicleDetails } from '@store/tests/vehicle-details/vehicle-details.reducer';
+import { getGearboxCategory } from '@store/tests/vehicle-details/vehicle-details.selector';
+import { getTestSummary } from '@store/tests/test-summary/test-summary.reducer';
+import { getD255, isDebriefWitnessed } from '@store/tests/test-summary/test-summary.selector';
+import { getCommunicationPreference } from '@store/tests/communication-preferences/communication-preferences.reducer';
+import { getConductedLanguage } from '@store/tests/communication-preferences/communication-preferences.selector';
+import { getTestData } from '@store/tests/test-data/cat-b/test-data.reducer';
+import { hasEyesightTestGotSeriousFault } from '@store/tests/test-data/cat-b/test-data.cat-b.selector';
+import { GearboxCategory } from '@dvsa/mes-test-schema/categories/common';
+import {
+  PassCertificateNumberChanged,
+  ProvisionalLicenseNotReceived,
+  ProvisionalLicenseReceived,
+} from '@store/tests/pass-completion/pass-completion.actions';
+import { GearboxCategoryChanged } from '@store/tests/vehicle-details/vehicle-details.actions';
+import {
+  D255No, D255Yes, DebriefUnWitnessed, DebriefWitnessed,
+} from '@store/tests/test-summary/test-summary.actions';
+import {
+  CandidateChoseToProceedWithTestInEnglish,
+  CandidateChoseToProceedWithTestInWelsh,
+} from '@store/tests/communication-preferences/communication-preferences.actions';
 
 export interface CommonPassFinalisationPageState {
   candidateName$: Observable<string>;
+  candidateUntitledName$: Observable<string>;
+  candidateDriverNumber$: Observable<string>;
+  testOutcomeText$: Observable<string>;
+  applicationNumber$: Observable<string>;
+  provisionalLicense$: Observable<boolean>;
+  passCertificateNumber$: Observable<string>;
+  transmission$: Observable<GearboxCategory>;
+  d255$: Observable<boolean>;
+  debriefWitnessed$: Observable<boolean>;
+  conductedLanguage$: Observable<string>;
+  eyesightTestFailed$: Observable<boolean>;
 }
 
 export abstract class PassFinalisationPageComponent extends PracticeableBasePageComponent {
@@ -41,7 +89,86 @@ export abstract class PassFinalisationPageComponent extends PracticeableBasePage
         select(getCandidate),
         select(getCandidateName),
       ),
+      candidateUntitledName$: currentTest$.pipe(
+        select(getJournalData),
+        select(getCandidate),
+        select(getUntitledCandidateName),
+      ),
+      candidateDriverNumber$: currentTest$.pipe(
+        select(getJournalData),
+        select(getCandidate),
+        select(getCandidateDriverNumber),
+        map(formatDriverNumber),
+      ),
+      testOutcomeText$: currentTest$.pipe(
+        select(getTestOutcomeText),
+      ),
+      applicationNumber$: currentTest$.pipe(
+        select(getJournalData),
+        select(getApplicationReference),
+        select(getApplicationNumber),
+      ),
+      provisionalLicense$: currentTest$.pipe(
+        select(getPassCompletion),
+        map(isProvisionalLicenseProvided),
+      ),
+      passCertificateNumber$: currentTest$.pipe(
+        select(getPassCompletion),
+        select(getPassCertificateNumber),
+      ),
+      transmission$: currentTest$.pipe(
+        select(getVehicleDetails),
+        select(getGearboxCategory),
+      ),
+      debriefWitnessed$: currentTest$.pipe(
+        select(getTestSummary),
+        select(isDebriefWitnessed),
+      ),
+      d255$: currentTest$.pipe(
+        select(getTestSummary),
+        select(getD255),
+      ),
+      conductedLanguage$: currentTest$.pipe(
+        select(getCommunicationPreference),
+        select(getConductedLanguage),
+      ),
+      eyesightTestFailed$: currentTest$.pipe(
+        select(getTestData),
+        select(hasEyesightTestGotSeriousFault),
+      ),
     };
+  }
+
+  provisionalLicenseReceived(): void {
+    this.store$.dispatch(ProvisionalLicenseReceived());
+  }
+
+  provisionalLicenseNotReceived(): void {
+    this.store$.dispatch(ProvisionalLicenseNotReceived());
+  }
+
+  transmissionChanged(transmission: GearboxCategory): void {
+    this.store$.dispatch(GearboxCategoryChanged(transmission));
+  }
+
+  passCertificateNumberChanged(passCertificateNumber: string): void {
+    this.store$.dispatch(PassCertificateNumberChanged(passCertificateNumber));
+  }
+
+  d255Changed(d255: boolean): void {
+    this.store$.dispatch(d255 ? D255Yes() : D255No());
+  }
+
+  debriefWitnessedChanged(debriefWitnessed: boolean) {
+    this.store$.dispatch(debriefWitnessed ? DebriefWitnessed() : DebriefUnWitnessed());
+  }
+
+  isWelshChanged(isWelsh: boolean) {
+    this.store$.dispatch(
+      isWelsh
+        ? CandidateChoseToProceedWithTestInWelsh('Cymraeg')
+        : CandidateChoseToProceedWithTestInEnglish('English'),
+    );
   }
 
 }
